feat: code structure refactoring
