+++ rs-react-app/src/pages/index.ts
@@ -1,1 +1,1 @@
-export * from './ui';
+export * from './main';
