+++ rs-react-app/src/pages/main/api/get-pokemon.ts
@@ -1,4 +1,4 @@
-import type { PokemonType } from '../interfaces';
+import type { PokemonType } from '@/shared';
 
 export const getPokemon = async (
   url: string
