+++ rs-react-app/src/pages/main/api/index.ts
@@ -1,0 +1,1 @@
+export * from './get-pokemon';
