+++ rs-react-app/src/pages/main/ui/index.ts
