+++ rs-react-app/src/pages/main/ui/search-section/index.ts
