+++ rs-react-app/src/pages/main/ui/search-section/search-input.tsx
@@ -1,4 +1,4 @@
-import type { SearchInputType } from '@/shared';
+import { getStorage, setStorage, type SearchInputType } from '@/shared';
 import React from 'react';
 import { type ChangeEvent } from 'react';
 
@@ -12,6 +12,12 @@ export class SearchInput extends React.Component<SearchInputType> {
     this.handleSubmit = this.handleSubmit.bind(this);
   }
 
+  componentDidMount(): void {
+    const storageValue = getStorage();
+    this.setState({ searchValue: storageValue });
+    if (storageValue) this.props.onSearch(storageValue);
+  }
+
   handleInputChange(event: ChangeEvent<HTMLInputElement>): void {
     if (event.target) this.setState({ searchValue: event.target.value });
   }
@@ -19,6 +25,7 @@ export class SearchInput extends React.Component<SearchInputType> {
   handleSubmit(event: { preventDefault: () => void }): void {
     event.preventDefault();
     const { searchValue } = this.state;
+    setStorage(searchValue);
     this.props.onSearch(searchValue);
   }
