+++ rs-react-app/src/pages/main/ui/table-section/index.ts
@@ -1,0 +1,1 @@
+export { Table } from './table';
