--- rs-react-app/src/pages/ui/index.ts
@@ -1,1 +1,0 @@
-export * from './main';
