--- rs-react-app/src/pages/ui/main/table-section/index.ts
@@ -1,1 +1,0 @@
-export { CardList } from './card-list';
