--- rs-react-app/src/shared/api/index.ts
@@ -1,1 +1,0 @@
-export { getPokemon } from './get-pokemon';
