+++ rs-react-app/src/shared/index.ts
@@ -1,4 +1,3 @@
-export * from './config';
-export * from './ui';
-export * from './api';
+export * from './error-boundary';
 export * from './interfaces';
+export * from './storage';
