--- rs-react-app/src/shared/ui/index.ts
@@ -1,1 +1,0 @@
-export { ErrorBoundary } from './error-boundary';
